testing

+++ src/vsr/models/VRT/modules/spynet.py
@@ -31,7 +31,7 @@ def flow_warp(x, flow, interp_mode='bilinear', padding_mode='zeros', align_corne
     n, _, h, w = x.size()
     # create mesh grid
     grid_y, grid_x = torch.meshgrid(torch.arange(0, h, dtype=x.dtype, device=x.device), torch.arange(0, w, dtype=x.dtype, device=x.device))
-    grid = torch.stack((grid_x, grid_y), 2).float()  # W(x), H(y), 2
+    grid = torch.stack((grid_x, grid_y), 2)  # W(x), H(y), 2
     grid.requires_grad = False
 
     vgrid = grid + flow
